refactor: extract database logic into separate module
